MK3/requirements.txt
fastapi>=0.118.3
uvicorn[standard]>=0.34.0
requests>=2.32.3
pydantic>=2.12.0
beautifulsoup4>=4.13.4
lxml>=6.0.2
trafilatura>=2.0.0
python-dotenv>=1.1.0